Create a separate pkg for test helper functions
