+++ pkg/host/internal/kernel/kernel_test.go
@@ -1,38 +1,20 @@
 package kernel
 
 import (
-	"os"
-	"path/filepath"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/consts"
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/host/types"
-	"github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/vars"
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/fakefilesystem"
+	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/helpers"
 )
-
-func assertFileContentsEquals(path, expectedContent string) {
-	d, err := os.ReadFile(filepath.Join(vars.FilesystemRoot, path))
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	ExpectWithOffset(1, string(d)).To(Equal(expectedContent))
-}
 
 var _ = Describe("Kernel", func() {
 	Context("Drivers", func() {
 		var (
 			k types.KernelInterface
 		)
-		configureFS := func(f *fakefilesystem.FS) {
-			var (
-				cleanFakeFs func()
-				err         error
-			)
-			vars.FilesystemRoot, cleanFakeFs, err = f.Use()
-			Expect(err).ToNot(HaveOccurred())
-			DeferCleanup(cleanFakeFs)
-		}
 		BeforeEach(func() {
 			k = New(nil)
 		})
@@ -41,11 +23,11 @@ var _ = Describe("Kernel", func() {
 				Expect(k.UnbindDriverByBusAndDevice(consts.BusPci, "unknown-dev")).NotTo(HaveOccurred())
 			})
 			It("known device, no driver", func() {
-				configureFS(&fakefilesystem.FS{Dirs: []string{"/sys/bus/pci/devices/0000:d8:00.0"}})
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{Dirs: []string{"/sys/bus/pci/devices/0000:d8:00.0"}})
 				Expect(k.Unbind("0000:d8:00.0")).NotTo(HaveOccurred())
 			})
 			It("has driver, succeed", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0",
 						"/sys/bus/pci/drivers/test-driver"},
@@ -56,10 +38,10 @@ var _ = Describe("Kernel", func() {
 				})
 				Expect(k.Unbind("0000:d8:00.0")).NotTo(HaveOccurred())
 				// check that echo to unbind path was done
-				assertFileContentsEquals("/sys/bus/pci/drivers/test-driver/unbind", "0000:d8:00.0")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/drivers/test-driver/unbind", "0000:d8:00.0")
 			})
 			It("has driver, failed to unbind", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0"},
 					Symlinks: map[string]string{
@@ -75,13 +57,13 @@ var _ = Describe("Kernel", func() {
 				Expect(driver).To(BeEmpty())
 			})
 			It("known device, no driver", func() {
-				configureFS(&fakefilesystem.FS{Dirs: []string{"/sys/bus/pci/devices/0000:d8:00.0"}})
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{Dirs: []string{"/sys/bus/pci/devices/0000:d8:00.0"}})
 				has, driver := k.HasDriver("0000:d8:00.0")
 				Expect(has).To(BeFalse())
 				Expect(driver).To(BeEmpty())
 			})
 			It("has driver", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0",
 						"/sys/bus/pci/drivers/test-driver"},
@@ -98,7 +80,7 @@ var _ = Describe("Kernel", func() {
 				Expect(k.BindDefaultDriver("unknown-dev")).To(HaveOccurred())
 			})
 			It("no driver", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0"},
 					Files: map[string][]byte{
@@ -106,10 +88,10 @@ var _ = Describe("Kernel", func() {
 				})
 				Expect(k.BindDefaultDriver("0000:d8:00.0")).NotTo(HaveOccurred())
 				// should probe driver for dev
-				assertFileContentsEquals("/sys/bus/pci/drivers_probe", "0000:d8:00.0")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/drivers_probe", "0000:d8:00.0")
 			})
 			It("already bind to default driver", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0"},
 					Symlinks: map[string]string{
@@ -118,7 +100,7 @@ var _ = Describe("Kernel", func() {
 				Expect(k.BindDefaultDriver("0000:d8:00.0")).NotTo(HaveOccurred())
 			})
 			It("bind to dpdk driver", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0",
 						"/sys/bus/pci/drivers/vfio-pci"},
@@ -130,9 +112,9 @@ var _ = Describe("Kernel", func() {
 				})
 				Expect(k.BindDefaultDriver("0000:d8:00.0")).NotTo(HaveOccurred())
 				// should unbind from dpdk driver
-				assertFileContentsEquals("/sys/bus/pci/drivers/vfio-pci/unbind", "0000:d8:00.0")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/drivers/vfio-pci/unbind", "0000:d8:00.0")
 				// should probe driver for dev
-				assertFileContentsEquals("/sys/bus/pci/drivers_probe", "0000:d8:00.0")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/drivers_probe", "0000:d8:00.0")
 			})
 		})
 		Context("BindDpdkDriver", func() {
@@ -140,7 +122,7 @@ var _ = Describe("Kernel", func() {
 				Expect(k.BindDpdkDriver("unknown-dev", "vfio-pci")).To(HaveOccurred())
 			})
 			It("no driver", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0",
 						"/sys/bus/pci/drivers/vfio-pci"},
@@ -149,10 +131,10 @@ var _ = Describe("Kernel", func() {
 				})
 				Expect(k.BindDpdkDriver("0000:d8:00.0", "vfio-pci")).NotTo(HaveOccurred())
 				// should reset driver override
-				assertFileContentsEquals("/sys/bus/pci/devices/0000:d8:00.0/driver_override", "\x00")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/devices/0000:d8:00.0/driver_override", "\x00")
 			})
 			It("already bind to required driver", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0"},
 					Symlinks: map[string]string{
@@ -161,7 +143,7 @@ var _ = Describe("Kernel", func() {
 				Expect(k.BindDpdkDriver("0000:d8:00.0", "vfio-pci")).NotTo(HaveOccurred())
 			})
 			It("bind to wrong driver", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0",
 						"/sys/bus/pci/drivers/test-driver",
@@ -175,12 +157,12 @@ var _ = Describe("Kernel", func() {
 				})
 				Expect(k.BindDpdkDriver("0000:d8:00.0", "vfio-pci")).NotTo(HaveOccurred())
 				// should unbind from driver1
-				assertFileContentsEquals("/sys/bus/pci/drivers/test-driver/unbind", "0000:d8:00.0")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/drivers/test-driver/unbind", "0000:d8:00.0")
 				// should bind to driver2
-				assertFileContentsEquals("/sys/bus/pci/drivers/vfio-pci/bind", "0000:d8:00.0")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/drivers/vfio-pci/bind", "0000:d8:00.0")
 			})
 			It("fail to bind", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0",
 						"/sys/bus/pci/drivers/test-driver"},
@@ -195,7 +177,7 @@ var _ = Describe("Kernel", func() {
 		})
 		Context("BindDriverByBusAndDevice", func() {
 			It("device doesn't support driver_override", func() {
-				configureFS(&fakefilesystem.FS{
+				helpers.GinkgoConfigureFakeFS(&fakefilesystem.FS{
 					Dirs: []string{
 						"/sys/bus/pci/devices/0000:d8:00.0",
 						"/sys/bus/pci/drivers/test-driver",
@@ -208,9 +190,9 @@ var _ = Describe("Kernel", func() {
 				})
 				Expect(k.BindDriverByBusAndDevice(consts.BusPci, "0000:d8:00.0", "vfio-pci")).NotTo(HaveOccurred())
 				// should unbind from driver1
-				assertFileContentsEquals("/sys/bus/pci/drivers/test-driver/unbind", "0000:d8:00.0")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/drivers/test-driver/unbind", "0000:d8:00.0")
 				// should bind to driver2
-				assertFileContentsEquals("/sys/bus/pci/drivers/vfio-pci/bind", "0000:d8:00.0")
+				helpers.GinkgoAssertFileContentsEquals("/sys/bus/pci/drivers/vfio-pci/bind", "0000:d8:00.0")
 			})
 		})
 	})
